// src/api/model/bankGet.ts
/**
 * Financial Management API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BankGet { 
    /**
     * Id of the created resource
     */
    readonly id: string;
    name: string;
    address?: string;
    phone?: string;
    readonly swiftCode?: string;
}